requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
ecdsa==0.19.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
sqlalchemy==2.0.38
starlette==0.45.3
typing-extensions==4.12.2
uvicorn==0.34.0
pytest
httpx
pytest-asyncio